client/pyproject.toml
[project]
name = "remote-control-client"
version = "1.0.0"
description = "Remote desktop control system - Client agent"
requires-python = ">=3.11"
dependencies = [
    "aiortc>=1.6.0",
    "av>=11.0.0",
    "pywayland>=0.4.17",
    "dbus-next>=0.2.3",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "pydantic>=2.5.0",
    "numpy>=2.0.0",
    "Pillow>=10.0.0",
    "python-xlib>=0.33",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
